Fixed a huge bug in CoMs calculation
The previous algorithm was taking the smallest and biggest number inside the molecule (#M) block, and then retrieve the molecule in the .xyz file. This way does not work all the times, e.g. if you scramble the atoms number and they are not following each molecule from the largest to the smallest